Handle generics in NativeScript derive and `#[methods]`
The derive macros are now aware of both lifetime and type parameters.
`'static` bounds are added to both kinds, which improves the error
message shown when a user tries to put a lifetime into a NativeClass.

This behavior is chosen because unlike type parameters, lifetime
parameters aren't very useful in this context: only `'static` would be
supported anyway, and any attempted usage is hightly likely to have
originated from user error.